Fix for azure ping issue.
