+++ src/Host/Client/Impl/Microsoft.R.Host.Client.csproj
@@ -78,6 +78,7 @@
     <Compile Include="Definitions\IConsole.cs" />
     <Compile Include="Definitions\IRSessionSwitchBrokerTransaction.cs" />
     <Compile Include="Extensions\AboutHostExtensions.cs" />
+    <Compile Include="Extensions\UriExtensions.cs" />
     <Compile Include="Host\BrokerConnectionInfo.cs" />
     <Compile Include="Host\HostConnectionInfo.cs" />
     <Compile Include="Host\LocalCredentialsDecorator.cs" />
